Move edge-module out of __init__.py
